tests/test_minesweeper.py
import unittest
from unittest.mock import patch, MagicMock
import pygame
from random import randrange
from ctypes import windll

# Assuming the Minesweeper game code is in a file named minesweeper.py and we have imported everything needed
from minesweeper import Block, initialize_blocks, place_bombs, get_warning_count, find_path, handle_loss, handle_win, reset_game

class MinesweeperTestCase(unittest.TestCase):

    def setUp(self):
        # Initialize Pygame and create a screen to avoid errors
        pygame.init()
        self.screen = pygame.display.set_mode((350, 400))
        # Create blocks fixture
        self.blocks = []
        for x in range(0, 9):
            for y in range(0, 9):
                block = Block(pygame.Rect(x*34+20, y*34+20, 34, 34))
                block.index = len(self.blocks)
                block.x = x
                block.y = y
                self.blocks.append(block)
        self.total_bombs = 0

    def tearDown(self):
        pygame.quit()

    def test_initialize_blocks(self):
        initialize_blocks()
        self.assertEqual(len(self.blocks), 81, "There should be 81 blocks initialized.")

    @patch('minesweeper.randrange', side_effect=[0, 1, 2, 3, 4, 5, 6, 7, 8, 9])
    def test_place_bombs(self, mock_randrange):
        place_bombs()
        bomb_count = sum(1 for block in self.blocks if block.is_bomb)
        self.assertEqual(bomb_count, 10, "There should be exactly 10 bombs placed.")
        self.assertTrue(all(block.is_bomb for block in self.blocks[:10]), "First 10 blocks should be bombs due to mocked randrange.")

    def test_get_warning_count(self):
        self.blocks[0].is_bomb = True
        self.blocks[1].is_bomb = True
        warning_count = get_warning_count(self.blocks[10])
        self.assertEqual(warning_count, 2, "Warning count should be 2 when there are 2 bombs in adjacent blocks.")

    def test_find_path(self):
        self.blocks[0].flag_status = 0
        find_path(self.blocks[0])
        self.assertTrue(all(block.flag_status in (0, 1, 2, 3, 4, 5, 6, 7, 8) for block in self.blocks), "All blocks should be either revealed or have a warning count.")

    @patch('minesweeper.showMessageBox', return_value=1)
    @patch('minesweeper.pygame.quit')
    @patch('minesweeper.sys.exit')
    def test_handle_loss(self, mock_exit, mock_quit, mock_messagebox):
        handle_loss()
        self.assertTrue(mock_quit.called, "Pygame quit should be called on loss.")
        self.assertTrue(mock_exit.called, "System exit should be called on loss.")

    @patch('minesweeper.showMessageBox', return_value=1)
    @patch('minesweeper.pygame.quit')
    @patch('minesweeper.sys.exit')
    def test_handle_win(self, mock_exit, mock_quit, mock_messagebox):
        handle_win()
        self.assertTrue(mock_quit.called, "Pygame quit should be called on win.")
        self.assertTrue(mock_exit.called, "System exit should be called on win.")

    @patch('minesweeper.pygame.time.get_ticks', return_value=0)
    def test_reset_game(self, mock_ticks):
        reset_game()
        self.assertEqual(len(self.blocks), 81, "Blocks should be reinitialized on game reset.")
        bomb_count = sum(1 for block in self.blocks if block.is_bomb)
        self.assertEqual(bomb_count, 10, "There should be exactly 10 bombs placed after reset.")

if __name__ == "__main__":
    unittest.main()
